+++ Week2/huieunKim/src/components/Hobby.scss
@@ -1,6 +1,3 @@
-@use '../styles/variables' as v;
-@use '../styles/mixin' as m;
-
 .hobby-section {
     width: 100%;
     max-width: 1211px;
@@ -14,32 +11,4 @@
     grid-template-columns: repeat(3, 1fr);
     gap: 20px;
     margin-top: 40px;
-}
-
-.hobby-item {
-    background-color: v.$primary-color;
-    border-radius: 16px;
-    overflow: hidden;
-    display: flex;
-    flex-direction: column;
-
-    .hobby-label {
-        padding: 15px;
-        color: v.$white;
-        font-weight: bold;
-    }
-
-    .hobby-item-img-wrapper {
-        background-color: v.$white;
-        padding: 40px;
-        @include m.flex-center;
-        height: 300px;
-
-        img {
-            width: 100%;
-            height: 100%;
-            object-fit: cover;
-            border-radius: 8px;
-        }
-    }
 }
