+++ Week2/huieunKim/src/components/HobbyItem.scss
@@ -1,0 +1,30 @@
+@use '../styles/variables' as v;
+@use '../styles/mixin' as m;
+
+.hobby-item {
+    background-color: v.$primary-color;
+    border-radius: 16px;
+    overflow: hidden;
+    display: flex;
+    flex-direction: column;
+
+    .hobby-label {
+        padding: 15px;
+        color: v.$white;
+        font-weight: bold;
+    }
+
+    .hobby-item-img-wrapper {
+        background-color: v.$white;
+        padding: 40px;
+        @include m.flex-center;
+        height: 300px;
+
+        img {
+            width: 100%;
+            height: 100%;
+            object-fit: cover;
+            border-radius: 8px;
+        }
+    }
+}
