+++ Week2/huieunKim/src/styles/global.scss
@@ -12,7 +12,7 @@ html {
 
 body {
     font-family: 'Inter', sans-serif;
-    background-color: $bg-color;
+    background-color: v.$bg-color;
     width: 100%;
     overflow-x: hidden;
 }
